+++ app/services/activities/import_from_csv.rb
@@ -128,6 +128,14 @@ module Activities
         @activity.organisation = @organisation
         @activity.reporting_organisation = @organisation
 
+        beis = Organisation.find_by(service_owner: true)
+        @activity.funding_organisation_name = beis.name
+        @activity.funding_organisation_reference = beis.iati_reference
+        @activity.funding_organisation_type = beis.organisation_type
+        @activity.accountable_organisation_name = beis.name
+        @activity.accountable_organisation_reference = beis.iati_reference
+        @activity.accountable_organisation_type = beis.organisation_type
+
         @activity.assign_attributes(@converter.to_h)
         @activity.level = calculate_level
         @activity.cache_roda_identifier